created localhost with Python
I just learned how to make python 3.4 a webserver on Windows 10. It took me only 2 hrs of trial and error until finally I realized that I had to "custom"  install python in C:\python37 and check the box that says "add environment variable" or something and then follow the tutorial on youtube: "How to make python 3.4 a webserver on Windows"!