complete "check balance" project
It was so hard that I had to google for the solution. I guess the problem was in the code requirements of the website. Coding is a flexible thing and it can be hard when u try to do it only one way. I tried to do it my way but it didn't work. Then, I learned a lesson from the solution. You don't have to put curly brackets "{ -- }" after the last "else" element.